Update httpurllib.py
For HTTPConnection.request(method, url[, body[, headers]]) in Python2.7, the body should be a string of data to send. But in Python3.7, which is the current function is based on, the body could be a string, bytes, or even files, depending on the contect. As there might be some unnormal symbols in the idf or gbXML file, Haopeng perviously used bytes to send.